FEAT:Highlighting works in mermaid and JSON when clicking node/edge in diagram
